+++ offworld/src/com/bitwaffle/offworld/interfaces/Health.java
@@ -7,11 +7,11 @@ package com.bitwaffle.offworld.interfaces;
  */
 public interface Health {
 	/** @return Current health of object */
-	public int currentHealth();
+	public float currentHealth();
 	
 	/** @param amount Amount to hurt object */
-	public void hurt(int amount);
+	public void hurt(float amount);
 	
 	/** @param amount Amount to heal object */
-	public void heal(int amount);
+	public void heal(float amount);
 }
